Update index.html
placing the title of the FAQ section inside a circle.  that circle could then be put over the line between the two slides.  I think that would look pretty cool (kind of like a sealed envelope).  I also think we need to decide on a background color for the slides.  black is extremely limiting.